simplify Box constructor
Remove poorly implemented feature of being able to create a new box like
another box, but with a few different parameters. If the feature is needed,
it should be reimplemented as part of of a copy or clone method.

Also, make products and byproducts a property to simplify produce() and
Box.finalize().